top level components from the base router are 'views'
