+++ app/components/MessagingBar.tsx
@@ -5,7 +5,7 @@ import { useHistory, useParams } from "react-router-dom";
 import { selectSessions, stubSession } from "../reducers/sessions";
 import { addFamiliar, selectFamiliars } from "../reducers/familiars";
 import { selectUser } from "../reducers/user";
-import { switchFX } from "../views/Base";
+import { switchFX } from "../constants/sounds";
 import { routes } from "../constants/routes";
 import { FamiliarButton } from "./FamiliarButton";
 import { UserSearchBar } from "./UserSearchBar";
