+++ app/constants/sounds.ts
@@ -1,0 +1,5 @@
+export const switchFX = new Audio("assets/sounds/switch_005.ogg");
+switchFX.load();
+
+export const errorFX = new Audio("assets/sounds/error_008.ogg");
+errorFX.load();
