+++ app/views/Base.tsx
@@ -1,27 +1,21 @@
 import React from "react";
-import { Switch, Route, Redirect } from "react-router-dom";
+import { Switch, Route } from "react-router-dom";
 import { routes } from "../constants/routes";
 import App from "../views/App";
 import Register from "../views/Register";
 import Settings from "../views/Settings";
-import { ClientLauncher } from "../components/ClientLauncher";
-import CreateServer from "../components/CreateServer";
+import { ClientLauncher } from "../views/ClientLauncher";
 import Messaging from "./Messaging";
 import { Server } from "./Server";
 import { KeyGaurdian } from "../utils/KeyGaurdian";
 import Store from "electron-store";
-import { Logout } from "../components/Logout";
-import { LoginForm } from "../components/LoginForm";
-import { IdentityPicker } from "../components/IdentityPicker";
+import { Logout } from "../views/Logout";
+import { Login } from "../views/Login";
 import { TitleBar } from "../components/TitleBar";
+import { Create } from "./Create";
+import { Home } from "./Home";
 
 export const gaurdian = new KeyGaurdian();
-
-export const switchFX = new Audio("assets/sounds/switch_005.ogg");
-switchFX.load();
-
-export const errorFX = new Audio("assets/sounds/error_008.ogg");
-errorFX.load();
 
 export const dataStore = new Store();
 
@@ -36,39 +30,18 @@ export default function Base(): JSX.Element {
                 />
                 <Route
                     path={routes.SERVERS + "/:serverID?/:channelID?"}
-                    render={({ match }) => <Server match={match} />}
+                    render={() => <Server />}
                 />
                 <Route path={routes.REGISTER} render={() => <Register />} />
                 <Route path={routes.SETTINGS} render={() => <Settings />} />
                 <Route path={routes.LAUNCH} render={() => <ClientLauncher />} />
                 <Route
                     path={routes.CREATE + "/:resourceType?"}
-                    render={({ match }) => {
-                        switch (match.params.resourceType) {
-                            case "server":
-                                return <CreateServer />;
-                            default:
-                                console.warn(
-                                    "Unsupported resource type for create route" +
-                                        match.params.resourceType
-                                );
-                                return;
-                        }
-                    }}
+                    render={() => <Create />}
                 />
-                <Route path={routes.LOGIN} render={() => <LoginForm />} />
+                <Route path={routes.LOGIN} render={() => <Login />} />
                 <Route path={routes.LOGOUT} render={() => <Logout />} />
-                <Route
-                    exact
-                    path={routes.HOME}
-                    render={() => {
-                        if (gaurdian.hasKey()) {
-                            return <Redirect to={routes.LAUNCH} />;
-                        } else {
-                            return <IdentityPicker />;
-                        }
-                    }}
-                />
+                <Route exact path={routes.HOME} render={() => <Home />} />
             </Switch>
         </App>
     );
